+++ tencentcloud/ccc/v20200210/models.go
@@ -2250,7 +2250,7 @@ type PhoneNumBuyInfo struct {
 	// 截至时间戳
 	EndTime *int64 `json:"EndTime,omitempty" name:"EndTime"`
 
-	// 号码状态，1正常|2停用
+	// 号码状态，1正常|2欠费停用|4管理员停用|5违规停用
 	State *int64 `json:"State,omitempty" name:"State"`
 }