+++ tencentcloud/cvm/v20170312/errors.go
@@ -50,6 +50,9 @@ const (
 	// 查询退换价格失败，找不到付款订单，请检查设备 `ins-xxxxxxx` 是否已过期。
 	FAILEDOPERATION_INQUIRYREFUNDPRICEFAILED = "FailedOperation.InquiryRefundPriceFailed"
 
+	// 镜像状态繁忙，请稍后重试。
+	FAILEDOPERATION_INVALIDIMAGESTATE = "FailedOperation.InvalidImageState"
+
 	// 请求不支持`EMR`的实例`ins-xxxxxxxx`。
 	FAILEDOPERATION_INVALIDINSTANCEAPPLICATIONROLEEMR = "FailedOperation.InvalidInstanceApplicationRoleEmr"
 
@@ -724,6 +727,9 @@ const (
 
 	// 云服务器绑定了弹性网卡，请解绑弹性网卡后再切换私有网络。
 	UNSUPPORTEDOPERATION_ELASTICNETWORKINTERFACE = "UnsupportedOperation.ElasticNetworkInterface"
+
+	// 不支持加密镜像。
+	UNSUPPORTEDOPERATION_ENCRYPTEDIMAGESNOTSUPPORTED = "UnsupportedOperation.EncryptedImagesNotSupported"
 
 	// 异构机型不支持跨机型调整。
 	UNSUPPORTEDOPERATION_HETEROGENEOUSCHANGEINSTANCEFAMILY = "UnsupportedOperation.HeterogeneousChangeInstanceFamily"